Issue #202: Reduce the file length of InputMutationTests
