+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests/DiffMutationTests.diff
@@ -1,6 +1,6 @@
 --- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests/InputMutationTests.java
 +++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests/OutputMutationTests.java
-@@ -9,234 +9,158 @@
+@@ -9,92 +9,65 @@
  
  package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests;
  
@@ -114,189 +114,4 @@
 +            }
          }
      }
- 
-     void sameLineStatements(int x) {
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1:      { doSomething("same");
-+            case 1 -> { doSomething("same");
-                 doSomething("next");
--                break;
-             }
-         }
-     }
- 
-     void blankLinesCase(int x) {
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
- 
--            case 1: {
-+            case 1 -> {
-                 doSomething("blankCase 1");
-                 doSomething("blankCase 2");
--                break;
-             }
-         }
-     }
- 
-     String returnOnlySwitch(int id) {
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (id) {
--            case 1:
--                return "REQUEST";
--            case 2:
--                return "REPLY";
--            default:
--                throw new IllegalArgumentException("Unknown id " + id);
--        }
-+        return switch (id) {
-+            case 1 -> "REQUEST";
-+            case 2 -> "REPLY";
-+            default -> throw new IllegalArgumentException("Unknown id " + id);
-+        };
-     }
- 
-     String multiStatementYield(int id) {
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (id) {
--            case 1:
-+        return switch (id) {
-+            case 1 -> {
-                 System.out.println("log");
--                return "one";
--            case 2:
--                return "two";
--            default:
--                throw new RuntimeException();
--        }
-+                yield "one";
-+            }
-+            case 2 -> "two";
-+            default -> throw new RuntimeException();
-+        };
-     }
- 
-     void assignmentSwitch(int x) {
--        String s;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1:
--                s = "one";
--                break;
--            case 2:
-+        String s = switch (x) {
-+            case 1 -> "one";
-+            case 2 -> {
-                 System.out.println("two");
--                s = "two";
--                break;
--            default:
--                s = "default";
--        }
-+                yield "two";
-+            }
-+            default -> "default";
-+        };
-     }
- 
-     void assignmentSwitchWithThrow(int x) {
--        String s;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1:
--                s = "one";
--                break;
--            default:
--                throw new RuntimeException("bad");
--        }
-+        String s = switch (x) {
-+            case 1 -> "one";
-+            default -> throw new RuntimeException("bad");
-+        };
-     }
- 
-     void assignmentSwitchFallthrough(int x) {
--        int val;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1:
--            case 2:
--                val = 10;
--                break;
--            case 3:
--                val = 30;
--                break;
--            default:
--                val = 0;
--                break;
--        }
-+        int val = switch (x) {
-+            case 1, 2 -> 10;
-+            case 3 -> 30;
-+            default -> 0;
-+        };
-     }
- 
-     void assignmentSwitchBlock(int x) {
--        String s;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1: {
--                s = "one";
--                break;
--            }
--            case 2: {
-+        String s = switch (x) {
-+            case 1 -> "one";
-+            case 2 -> {
-                 System.out.println("two");
--                s = "two";
--                break;
-+                yield "two";
-             }
--            default: {
--                s = "default";
--                break;
--            }
--        }
-+            default -> "default";
-+        };
-     }
- 
-     void assignmentSwitchBlockMulti(int x) {
--        String s;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1: {
-+        String s = switch (x) {
-+            case 1 -> {
-                 System.out.println("log1");
-                 System.out.println("log2");
--                s = "one";
--                break;
-+                yield "one";
-             }
--            default: {
--                s = "default";
--                break;
--            }
--        }
-+            default -> "default";
-+        };
-     }
- 
-     void assignmentSwitchInitialized(int x) {
-         String s = "init";
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1:
--                s = "one";
--                break;
--            default:
--                s = "default";
--                break;
-+            case 1 -> s = "one";
-+            default -> s = "default";
-         }
-     }
- 
+ }
