+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests2/DiffMutationTests2.diff
@@ -1,0 +1,133 @@
+--- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests2/InputMutationTests2.java
++++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests2/OutputMutationTests2.java
+@@ -9,99 +9,70 @@
+ 
+ package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests2;
+ 
+-public class InputMutationTests2 {
++public class OutputMutationTests2 {
+ 
+     void doSomething(String p) {}
+ 
+     void sameLineStatements(int x) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+         switch (x) {
+-            case 1:      { doSomething("same");
++            case 1 -> { doSomething("same");
+                 doSomething("next");
+-                break;
+             }
+         }
+     }
+ 
+     void blankLinesCase(int x) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+         switch (x) {
+ 
+-            case 1: {
++            case 1 -> {
+                 doSomething("blankCase 1");
+                 doSomething("blankCase 2");
+-                break;
+             }
+         }
+     }
+ 
+     String returnOnlySwitch(int id) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (id) {
+-            case 1:
+-                return "REQUEST";
+-            case 2:
+-                return "REPLY";
+-            default:
+-                throw new IllegalArgumentException("Unknown id " + id);
+-        }
++        return switch (id) {
++            case 1 -> "REQUEST";
++            case 2 -> "REPLY";
++            default -> throw new IllegalArgumentException("Unknown id " + id);
++        };
+     }
+ 
+     String multiStatementYield(int id) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (id) {
+-            case 1:
++        return switch (id) {
++            case 1 -> {
+                 System.out.println("log");
+-                return "one";
+-            case 2:
+-                return "two";
+-            default:
+-                throw new RuntimeException();
+-        }
++                yield "one";
++            }
++            case 2 -> "two";
++            default -> throw new RuntimeException();
++        };
+     }
+ 
+     void assignmentSwitch(int x) {
+-        String s;
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) {
+-            case 1:
+-                s = "one";
+-                break;
+-            case 2:
++        String s = switch (x) {
++            case 1 -> "one";
++            case 2 -> {
+                 System.out.println("two");
+-                s = "two";
+-                break;
+-            default:
+-                s = "default";
+-        }
++                yield "two";
++            }
++            default -> "default";
++        };
+     }
+ 
+     void assignmentSwitchWithThrow(int x) {
+-        String s;
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) {
+-            case 1:
+-                s = "one";
+-                break;
+-            default:
+-                throw new RuntimeException("bad");
+-        }
++        String s = switch (x) {
++            case 1 -> "one";
++            default -> throw new RuntimeException("bad");
++        };
+     }
+ 
+     void assignmentSwitchFallthrough(int x) {
+-        int val;
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) {
+-            case 1:
+-            case 2:
+-                val = 10;
+-                break;
+-            case 3:
+-                val = 30;
+-                break;
+-            default:
+-                val = 0;
+-                break;
+-        }
++        int val = switch (x) {
++            case 1, 2 -> 10;
++            case 3 -> 30;
++            default -> 0;
++        };
+     }
+ }
